rich type system (support for arrays)

+++ crates/parser/src/token_set.rs
@@ -14,10 +14,10 @@ use syntax::TokenKind;
 // https://github.com/rust-analyzer/rust-analyzer/blob/b73b321478d3b2a98d380eb79de717e01620c4e9/crates/parser/src/token_set.rs
 // (This implementation is also stolen lol)
 #[derive(Debug, Clone, Copy)]
-pub(crate) struct TokenSet(u32);
+pub(crate) struct TokenSet(u64);
 
 impl TokenSet {
-    pub(crate) const ALL: Self = Self(u32::MAX);
+    pub(crate) const ALL: Self = Self(u64::MAX);
     pub(crate) const NONE: Self = Self(0);
 
     pub(crate) const fn new<const LEN: usize>(kinds: [TokenKind; LEN]) -> Self {
@@ -41,8 +41,8 @@ impl TokenSet {
     }
 }
 
-const fn mask(kind: TokenKind) -> u32 {
-    1 << kind as u32
+const fn mask(kind: TokenKind) -> u64 {
+    1 << kind as u64
 }
